requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pyodbc
pandas
python-multipart
nltk
langchain
langchain-experimental
langchain-google-genai
redis
redisvl
pika
aio-pika
unstructured[pdf]
azure-search-documents
azure-identity
langchainhub